feat/ implement room card component
