+++ src/components/core/title/title.module.css
@@ -1,8 +1,12 @@
 .title {
-    text-align: center;
-    color: #DAC3AB;
-    background-color:#1c354e;
     margin: 0 auto;
-    width: 22%;
-    border-radius: 2px;
+    z-index: 3;
+    width: 30%;
+    font-size: 1.1rem;
+    text-align: center;
+    padding: 0.6rem;
+    background-color: #1c354e;
+    color: #f9f7f6;
+    font-weight: 400;
+    transform: translateY(50%); 
 }
