+++ src/components/product-card-simple/card.module.css
@@ -1,0 +1,39 @@
+.card {
+    color: #234465;
+    background: #eeeeee;
+    width: 100%;
+    height: auto;
+    border: 1px solid #234465;
+    margin-bottom: 1.6%;
+    border-radius: 4px;
+}
+
+.description {
+    word-wrap: break-word;
+    padding: 0.5%;
+    font-size: 18px;
+    /* border:1px solid blue; */
+    margin-bottom: 0.5%;
+    width: 92%;
+    display: inline-block;
+    vertical-align: top;
+}
+
+.card div {
+    /* border:1px solid red; */
+    height: auto;
+    text-align: right;
+    margin-bottom: 0.5%;
+}
+
+.card div span {
+    text-align: right;
+    /* border:1px solid green; */
+    padding-right: 1%;
+}
+
+.card img {
+    padding: 0.5%;
+    width: 42px;
+    height: 42px;
+}
