+++ src/components/product-card/card.module.css
@@ -1,39 +1,55 @@
-.card {
-    color: #234465;
-    background: #eeeeee;
+.room {
+    margin-top: 1%;
+    width: 30%;
+    background: rgb(238,238,238,0.81);
+}
+
+.room_image {
     width: 100%;
-    height: auto;
-    border: 1px solid #234465;
-    margin-bottom: 1.6%;
-    border-radius: 4px;
+    z-index: 0;
+    margin: 0 auto;
 }
 
-.description {
-    word-wrap: break-word;
-    padding: 0.5%;
-    font-size: 18px;
-    /* border:1px solid blue; */
-    margin-bottom: 0.5%;
-    width: 92%;
-    display: inline-block;
-    vertical-align: top;
+.room_name {
+    margin: 0 auto;
+    position: relative;
+    top: -3rem;
+    z-index: 1;
+    width: 80%;
+    font-size: 1.1rem;
+    text-align: center;
+    padding: 0.6rem;
+    background-color: #1c354e;
+    color: #f9f7f6;
+    font-weight: 400;
+    transform: translateY(50%);  
 }
 
-.card div {
-    /* border:1px solid red; */
-    height: auto;
-    text-align: right;
-    margin-bottom: 0.5%;
+.people {
+    margin: 0 auto;
+    padding: 1%;
+    display: grid;
+    grid-template-columns: 50% 50%;
 }
 
-.card div span {
-    text-align: right;
-    /* border:1px solid green; */
-    padding-right: 1%;
+.person {
+    font-size: 1rem;
+    margin: 1rem auto;
 }
 
-.card img {
-    padding: 0.5%;
-    width: 42px;
-    height: 42px;
+.room_btn {
+    background-color: forestgreen;
+    color:#f9f7f6;
+    border: none;
+    border-radius: 0;
+    font-size: 1.2rem;
+    text-transform: uppercase;
+    padding: 1.2rem;
+    cursor: pointer;
+    transform: all 0.2;
+    width: 100%;
+}
+
+.room_btn:hover {
+    background-color: #07ac72;
 }
