+++ src/pages/home-page/home.js
@@ -31,7 +31,6 @@ class Home extends React.Component {
     render() {
         return (
             <Common>
-                <Title title='Available Rooms' />
                 <> {this.state.data} </>
             </Common>
         )
