+++ src/pages/share-you-thoughts/share.js
@@ -3,7 +3,7 @@ import './share.module.css'
 
 import Common from '../../components/common/common';
 import Title from '../../components/core/title/title';
-import ProductCard from '../../components/product-card/card';
+import ProductCard from '../../components/product-card-simple/card';
 import productService from '../../services/product-service';
 import InformationPanel from '../../components/information-panel/information-panel';
 import DefinedButton from '../../components/core/button/button';
